help

+++ inc/display.h
@@ -4,4 +4,4 @@
 
 int displayinit();
 int displayrefresh();
-int draw(u16 *mem, u8 x, u8 y, u16 size);
+int draw(u8 *mem, u8 x, u8 y, u16 size);
